Add cli command for building and pushing app images to docker repo
Extracts existing functionality used in deploy into separate command.
